+++ crates/ferroid/src/generator/ulid/tests.rs
@@ -7,7 +7,6 @@ use std::collections::HashSet;
 use std::rc::Rc;
 use std::sync::{Arc, Mutex};
 use std::thread::scope;
-use std::u128;
 
 struct MockTime {
     millis: u128,
