+++ crates/ferroid/src/id/mod.rs
@@ -1,11 +1,11 @@
-mod id;
+mod interface;
 #[cfg(feature = "snowflake")]
 mod snowflake;
 mod to_u64;
 #[cfg(feature = "ulid")]
 mod ulid;
 
-pub use id::*;
+pub use interface::*;
 #[cfg(feature = "snowflake")]
 pub use snowflake::*;
 pub use to_u64::*;
